@@ -13,17 +13,26 @@
         }
     ],
      "require": {
-        "php": "~7.1",
-        "silex/silex": "~2.0",
-        "guzzle/guzzle": ">=3.8",
-        "symfony/process": "~3|~4|~5",
-        "jeremeamia/superclosure": "~2",
-        "lstrojny/hmmmath": ">=0.5.0"
-    },
+        "php": "~7.4|~8.0",
+        "symfony/process": "~4.4|~5",
+        "lstrojny/hmmmath": ">=0.5.0",
+        "slevomat/coding-standard": "^7.0",
+        "opis/closure": "^3.6",
+        "guzzlehttp/guzzle": "^7.3",
+        "psr/http-factory": "^1.0",
+        "guzzlehttp/psr7": "^1.8",
+        "http-interop/http-factory-guzzle": "^1.0",
+        "php-http/discovery": "^1.13",
+        "symfony/config": "~4.4|~5",
+        "symfony/http-kernel": "~4.4|~5",
+        "symfony/http-foundation": "~4.4|~5",
+        "symfony/routing": "~4.4|~5",
+        "symfony/dependency-injection": "~4.4|~5",
+        "symfony/framework-bundle": "~4.4|~5"
+     },
     "require-dev": {
-        "internations/kodierungsregelwerksammlung": "~0.23.0",
-        "internations/testing-component": "1.0.1",
-        "phpunit/phpunit": "^7"
+        "phpunit/phpunit": "^9",
+        "qossmic/deptrac-shim": "^0.13.0"
     },
     "autoload": {
         "psr-4": {"InterNations\\Component\\HttpMock\\": "src/"}
@@ -32,6 +41,13 @@
         "psr-4": {"InterNations\\Component\\HttpMock\\Tests\\": "tests/"}
     },
     "scripts": {
-        "tests": "phpunit"
+        "tests": "phpunit",
+        "coding-style": "phpcs",
+        "dependencies": "deptrac",
+        "checkup": [
+            "@composer tests",
+            "@composer coding-style",
+            "@composer dependencies"
+        ]
     }
 }
