@@ -1,17 +1,15 @@
 <?php
-namespace InterNations\Component\HttpMock;
+namespace InterNations\Component\HttpMock\Response;
 
-use InterNations\Component\HttpMock\Response\CallbackResponse;
-use SuperClosure\SerializableClosure;
+use InterNations\Component\HttpMock\MockBuilder;
 use Closure;
+use Opis\Closure\SerializableClosure;
+use Symfony\Component\HttpFoundation\Response;
 
-class ResponseBuilder
+final class ResponseBuilder
 {
-    /** @var MockBuilder */
-    private $mockBuilder;
-
-    /** @var CallbackResponse */
-    private $response;
+    private MockBuilder $mockBuilder;
+    private CallbackResponse $response;
 
     public function __construct(MockBuilder $mockBuilder)
     {
@@ -19,40 +17,40 @@ class ResponseBuilder
         $this->response = new CallbackResponse();
     }
 
-    public function statusCode($statusCode)
+    public function statusCode(int $statusCode): self
     {
         $this->response->setStatusCode($statusCode);
 
         return $this;
     }
 
-    public function body($body)
+    public function body(string $body): self
     {
         $this->response->setContent($body);
 
         return $this;
     }
 
-    public function callback(Closure $callback)
+    public function callback(Closure $callback): self
     {
         $this->response->setCallback(new SerializableClosure($callback));
 
         return $this;
     }
 
-    public function header($header, $value)
+    public function header(string $header, string $value): self
     {
         $this->response->headers->set($header, $value);
 
         return $this;
     }
 
-    public function end()
+    public function end(): MockBuilder
     {
         return $this->mockBuilder;
     }
 
-    public function getResponse()
+    public function getResponse(): Response
     {
         return $this->response;
     }
