@@ -2,17 +2,15 @@
 namespace InterNations\Component\HttpMock\Tests\Matcher;
 
 use InterNations\Component\HttpMock\Matcher\StringMatcher;
-use InterNations\Component\Testing\AbstractTestCase;
+use InterNations\Component\HttpMock\Tests\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 
-class StringMatcherTest extends AbstractTestCase
+class StringMatcherTest extends TestCase
 {
-    public function testConversionToString()
+    public function testConversionToString(): void
     {
         $matcher = new StringMatcher('0');
-        $matcher->setExtractor(static function() {
-           return 0;
-        });
+        $matcher->setExtractor(static fn () => 0);
         self::assertTrue($matcher->getMatcher()(new Request()));
     }
 }
