@@ -1,12 +1,10 @@
 <?php
 namespace InterNations\Component\HttpMock\Tests;
 
-use InterNations\Component\HttpMock\Expectation;
 use InterNations\Component\HttpMock\Matcher\ExtractorFactory;
 use InterNations\Component\HttpMock\Matcher\MatcherFactory;
 use InterNations\Component\HttpMock\MockBuilder;
-use InterNations\Component\HttpMock\Server;
-use PHPUnit\Framework\TestCase;
+use InterNations\Component\HttpMock\ServerProcess;
 use DateTime;
 use DateTimeZone;
 use InterNations\Component\HttpMock\Tests\Fixtures\Request as TestRequest;
@@ -18,30 +16,27 @@ use Symfony\Component\HttpFoundation\Request;
  */
 class MockBuilderIntegrationTest extends TestCase
 {
-    /** @var MockBuilder */
-    private $builder;
+    private MockBuilder $builder;
 
-    /** @var MatcherFactory */
-    private $matches;
+    private MatcherFactory $matches;
 
-    /** @var Server */
-    private $server;
+    private ServerProcess $server;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->matches = new MatcherFactory();
         $this->builder = new MockBuilder($this->matches, new ExtractorFactory());
-        $this->server = new Server(HTTP_MOCK_PORT, HTTP_MOCK_HOST);
+        $this->server = new ServerProcess(HTTP_MOCK_PORT, HTTP_MOCK_HOST);
         $this->server->start();
         $this->server->clean();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         $this->server->stop();
     }
 
-    public function testCreateExpectation()
+    public function testCreateExpectation(): void
     {
         $builder = $this->builder
             ->when()
@@ -53,15 +48,15 @@ class MockBuilderIntegrationTest extends TestCase
                 })
             ->then()
                 ->statusCode(401)
-                ->body('response body')
                 ->header('X-Foo', 'Bar')
+                ->body('response body')
             ->end();
 
-        $this->assertSame($this->builder, $builder);
+        self::assertSame($this->builder, $builder);
 
         $expectations = $this->builder->flushExpectations();
 
-        $this->assertCount(1, $expectations);
+        self::assertCount(1, $expectations);
         /** @var Expectation $expectation */
         $expectation = current($expectations);
 
@@ -72,31 +67,34 @@ class MockBuilderIntegrationTest extends TestCase
         $run = 0;
         $oldValue = ini_set('error_log', '/dev/null');
         foreach ($expectation->getMatcherClosures() as $closure) {
-            $this->assertTrue($closure($request));
+            self::assertTrue($closure($request));
 
             $unserializedClosure = unserialize(serialize($closure));
-            $this->assertTrue($unserializedClosure($request));
+            self::assertTrue($unserializedClosure($request));
 
             $run++;
         }
         ini_set('error_log', $oldValue);
-        $this->assertSame(3, $run);
+        self::assertSame(3, $run);
 
         $expectation->getResponse()->setDate(new DateTime('2012-11-10 09:08:07', new DateTimeZone('UTC')));
-        $response = "HTTP/1.0 401 Unauthorized\r\nCache-Control: no-cache, private\r\nDate:          Sat, 10 Nov 2012 09:08:07 GMT\r\nX-Foo:         Bar\r\n\r\nresponse body";
-        $this->assertSame($response, (string)$expectation->getResponse());
+        $response = "HTTP/1.0 401 Unauthorized\r\nCache-Control: no-cach%s\r\nDate:          Sat, 10 Nov 2012 09:08:07 GMT\r\nX-Foo:         Bar\r\n\r\nresponse body";
+        self::assertStringMatchesFormat($response, (string)$expectation->getResponse());
 
 
         $this->server->setUp($expectations);
 
         $client = $this->server->getClient();
 
-        $this->assertSame('response body', (string) $client->post('/foo')->send()->getBody());
+        self::assertSame(
+            'response body',
+            (string) $client->sendRequest($this->getRequestFactory()->createRequest('POST', '/foo'))->getBody()
+        );
 
-        $this->assertContains('CLOSURE MATCHER: POST /foo', $this->server->getErrorOutput());
+        self::assertNotFalse(strpos($this->server->getErrorOutput(), 'CLOSURE MATCHER: POST /foo'));
     }
 
-    public function testCreateTwoExpectationsAfterEachOther()
+    public function testCreateTwoExpectationsAfterEachOther(): void
     {
         $this->builder
             ->when()
@@ -118,13 +116,33 @@ class MockBuilderIntegrationTest extends TestCase
             ->end();
         $this->server->setUp($this->builder->flushExpectations());
 
-        $this->assertSame('POST 1', (string) $this->server->getClient()->post('/post-resource-1')->send()->getBody());
-        $this->assertSame('POST 2', (string) $this->server->getClient()->post('/post-resource-2')->send()->getBody());
-        $this->assertSame('POST 1', (string) $this->server->getClient()->post('/post-resource-1')->send()->getBody());
-        $this->assertSame('POST 2', (string) $this->server->getClient()->post('/post-resource-2')->send()->getBody());
+        self::assertSame(
+            'POST 1',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/post-resource-1')
+            )->getBody()
+        );
+        self::assertSame(
+            'POST 2',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/post-resource-2')
+            )->getBody()
+        );
+        self::assertSame(
+            'POST 1',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/post-resource-1')
+            )->getBody()
+        );
+        self::assertSame(
+            'POST 2',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/post-resource-2')
+            )->getBody()
+        );
     }
 
-    public function testCreateSuccessiveExpectationsOnSameWhen()
+    public function testCreateSuccessiveExpectationsOnSameWhen(): void
     {
       $this->builder
           ->first()
@@ -150,12 +168,27 @@ class MockBuilderIntegrationTest extends TestCase
 
       $this->server->setUp($this->builder->flushExpectations());
 
-      $this->assertSame('called once', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-      $this->assertSame('called twice', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-      $this->assertSame('called 3 times', (string) $this->server->getClient()->post('/resource')->send()->getBody());
+        self::assertSame(
+            'called once',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            'called twice',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            'called 3 times',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
     }
 
-    public function testCreateSuccessiveExpectationsWithAny()
+    public function testCreateSuccessiveExpectationsWithAny(): void
     {
         $this->builder
             ->first()
@@ -181,12 +214,12 @@ class MockBuilderIntegrationTest extends TestCase
 
         $this->server->setUp($this->builder->flushExpectations());
 
-        $this->assertSame('1', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('2', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('any', (string) $this->server->getClient()->post('/resource')->send()->getBody());
+        self::assertSame('1', (string) $this->server->getClient()->sendRequest($this->getRequestFactory()->createRequest('POST', '/resource'))->getBody());
+        self::assertSame('2', (string) $this->server->getClient()->sendRequest($this->getRequestFactory()->createRequest('POST', '/resource'))->getBody());
+        self::assertSame('any', (string) $this->server->getClient()->sendRequest($this->getRequestFactory()->createRequest('POST', '/resource'))->getBody());
     }
 
-    public function testCreateSuccessiveExpectationsInUnexpectedOrder()
+    public function testCreateSuccessiveExpectationsInUnexpectedOrder(): void
     {
         $this->builder
             ->second()
@@ -205,11 +238,11 @@ class MockBuilderIntegrationTest extends TestCase
 
         $this->server->setUp($this->builder->flushExpectations());
 
-        $this->assertSame('1', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('2', (string) $this->server->getClient()->post('/resource')->send()->getBody());
+        self::assertSame('1', (string) $this->server->getClient()->sendRequest($this->getRequestFactory()->createRequest('POST', '/resource'))->getBody());
+        self::assertSame('2', (string) $this->server->getClient()->sendRequest($this->getRequestFactory()->createRequest('POST', '/resource'))->getBody());
     }
 
-    public function testCreateSuccessiveExpectationsWithOnce()
+    public function testCreateSuccessiveExpectationsWithOnce(): void
     {
         $this->builder
             ->first()
@@ -235,10 +268,35 @@ class MockBuilderIntegrationTest extends TestCase
 
         $this->server->setUp($this->builder->flushExpectations());
 
-        $this->assertSame('1', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('2', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('twice', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('twice', (string) $this->server->getClient()->post('/resource')->send()->getBody());
-        $this->assertSame('Expectation not met', (string) $this->server->getClient()->post('/resource')->send()->getBody());
+        self::assertSame(
+            '1',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            '2',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            'twice',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            'twice',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
+        self::assertSame(
+            'No matching expectation found',
+            (string) $this->server->getClient()->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/resource')
+            )->getBody()
+        );
     }
 }
