@@ -3,19 +3,19 @@ namespace InterNations\Component\HttpMock\Matcher;
 
 use Closure;
 
-class MatcherFactory
+final class MatcherFactory
 {
-    public function regex($regex)
+    public function regex(string $regex): Matcher
     {
         return new RegexMatcher($regex);
     }
 
-    public function str($string)
+    public function str(string $string): Matcher
     {
         return new StringMatcher($string);
     }
 
-    public function closure(Closure $closure)
+    public function closure(Closure $closure): Matcher
     {
         return new ClosureMatcher($closure);
     }
