@@ -5,22 +5,25 @@ use Symfony\Component\HttpFoundation\Response;
 
 class CallbackResponse extends Response
 {
+    /** @var callable */
     private $callback;
 
-    public function setCallback(callable $callback)
+    public function setCallback(callable $callback): void
     {
         $this->callback = $callback;
     }
 
-    public function sendCallback()
+    public function sendCallback(): void
     {
-        if ($this->callback) {
-            $callback = $this->callback;
-            $callback($this);
+        if (!$this->callback) {
+            return;
         }
+
+        $callback = $this->callback;
+        $callback($this);
     }
 
-    public function send()
+    public function send(): void
     {
         $this->sendCallback();
         parent::send();
