@@ -1,16 +1,16 @@
 <?php
 namespace InterNations\Component\HttpMock\Tests\Fixtures;
 
-use Symfony\Component\HttpFoundation\Request as BaseRequest;
+use InterNations\Component\HttpMock\Request\SerializableRequest;
 
-class Request extends BaseRequest
+class Request extends SerializableRequest
 {
-    public function setRequestUri($requestUri)
+    public function setRequestUri(string $requestUri): void
     {
         $this->requestUri = $requestUri;
     }
 
-    public function setContent($content)
+    public function setContent(string $content): void
     {
         $this->content = $content;
     }
