@@ -1,49 +1,48 @@
 <?php
 namespace InterNations\Component\HttpMock\Tests\PHPUnit;
 
-use InterNations\Component\HttpMock\PHPUnit\HttpMockTrait;
-use InterNations\Component\Testing\AbstractTestCase;
+use InterNations\Component\HttpMock\PHPUnit\HttpMock;
+use InterNations\Component\HttpMock\Tests\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use PHPUnit\Framework\TestCase;
+use function http_build_query;
 
 /** @large */
-class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
+class HttpMockPHPUnitIntegrationTest extends TestCase
 {
-    use HttpMockTrait;
+    use HttpMock;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         static::setUpHttpMockBeforeClass();
     }
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
         static::tearDownHttpMockAfterClass();
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->setUpHttpMock();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         $this->tearDownHttpMock();
     }
 
-    public static function getPaths()
+    /** @return array<array{0:string}> */
+    public static function getPaths(): array
     {
         return [
-            [
-                '/foo',
-                '/bar',
-            ]
+            ['/foo'],
+            ['/bar'],
         ];
     }
 
     /** @dataProvider getPaths */
-    public function testSimpleRequest($path)
+    public function testSimpleRequest(string $path): void
     {
         $this->http->mock
             ->when()
@@ -53,33 +52,43 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
             ->end();
         $this->http->setUp();
 
-        $this->assertSame($path . ' body', (string) $this->http->client->get($path)->send()->getBody());
+        self::assertSame(
+            $path . ' body',
+            (string) $this->http->client->sendRequest(
+                $this->getRequestFactory()->createRequest('GET', $path)
+            )->getBody()
+        );
 
         $request = $this->http->requests->latest();
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
         $request = $this->http->requests->last();
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
         $request = $this->http->requests->first();
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
         $request = $this->http->requests->at(0);
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
         $request = $this->http->requests->pop();
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
-        $this->assertSame($path . ' body', (string) $this->http->client->get($path)->send()->getBody());
+        self::assertSame(
+            $path . ' body',
+            (string) $this->http->client->sendRequest(
+                $this->getRequestFactory()->createRequest('GET', $path)
+            )->getBody()
+        );
 
         $request = $this->http->requests->shift();
-        $this->assertSame('GET', $request->getMethod());
-        $this->assertSame($path, $request->getPath());
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame($path, $request->getRequestUri());
 
         $this->expectException('UnexpectedValueException');
 
@@ -87,46 +96,46 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
         $this->http->requests->pop();
     }
 
-    public function testErrorLogOutput()
+    public function testErrorLogOutput(): void
     {
         $this->http->mock
             ->when()
-                ->callback(static function () {error_log('error output');})
+                ->callback(static function (): void {error_log('error output');})
             ->then()
             ->end();
         $this->http->setUp();
 
-        $this->http->client->get('/foo')->send();
+        $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/foo'));
 
         // Should fail during tear down as we have an error_log() on the server side
         try {
             $this->tearDown();
-            $this->fail('Exception expected');
+            self::fail('Exception expected');
         } catch (\Exception $e) {
-            $this->assertContains('HTTP mock server standard error output should be empty', $e->getMessage());
+            self::assertNotFalse(strpos($e->getMessage(), 'HTTP mock server standard error output should be empty'));
         }
     }
 
-    public function testFailedRequest()
+    public function testFailedRequest(): void
     {
-        $response = $this->http->client->get('/foo')->send();
-        $this->assertSame(404, $response->getStatusCode());
-        $this->assertSame('No matching expectation found', (string) $response->getBody());
+        $response = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/foo'));
+        self::assertSame(404, $response->getStatusCode());
+        self::assertSame('No matching expectation found', (string) $response->getBody());
     }
 
-    public function testStopServer()
+    public function testStopServer(): void
     {
         $this->http->server->stop();
     }
 
     /** @depends testStopServer */
-    public function testHttpServerIsRestartedIfATestStopsIt()
+    public function testHttpServerIsRestartedIfATestStopsIt(): void
     {
-        $response = $this->http->client->get('/')->send();
-        $this->assertSame(404, $response->getStatusCode());
+        $response = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/foo'));
+        self::assertSame(404, $response->getStatusCode());
     }
 
-    public function testLimitDurationOfAResponse()
+    public function testLimitDurationOfAResponse(): void
     {
         $this->http->mock
             ->once()
@@ -136,11 +145,11 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->body('POST METHOD')
             ->end();
         $this->http->setUp();
-        $firstResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $firstResponse->getStatusCode());
-        $secondResponse = $this->http->client->post('/')->send();
-        $this->assertSame(410, $secondResponse->getStatusCode());
-        $this->assertSame('Expectation not met', $secondResponse->getBody(true));
+        $firstResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $firstResponse->getStatusCode());
+        $secondResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(404, $secondResponse->getStatusCode());
+        self::assertSame('No matching expectation found', (string) $secondResponse->getBody());
 
         $this->http->mock
             ->exactly(2)
@@ -150,13 +159,13 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->body('POST METHOD')
             ->end();
         $this->http->setUp();
-        $firstResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $firstResponse->getStatusCode());
-        $secondResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $secondResponse->getStatusCode());
-        $thirdResponse = $this->http->client->post('/')->send();
-        $this->assertSame(410, $thirdResponse->getStatusCode());
-        $this->assertSame('Expectation not met', $thirdResponse->getBody(true));
+        $firstResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $firstResponse->getStatusCode());
+        $secondResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $secondResponse->getStatusCode());
+        $thirdResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(404, $thirdResponse->getStatusCode());
+        self::assertSame('No matching expectation found', (string) $thirdResponse->getBody());
 
         $this->http->mock
             ->any()
@@ -166,27 +175,32 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->body('POST METHOD')
             ->end();
         $this->http->setUp();
-        $firstResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $firstResponse->getStatusCode());
-        $secondResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $secondResponse->getStatusCode());
-        $thirdResponse = $this->http->client->post('/')->send();
-        $this->assertSame(200, $thirdResponse->getStatusCode());
+        $firstResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $firstResponse->getStatusCode());
+        $secondResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $secondResponse->getStatusCode());
+        $thirdResponse = $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'));
+        self::assertSame(200, $thirdResponse->getStatusCode());
     }
 
-    public function testCallbackOnResponse()
+    public function testCallbackOnResponse(): void
     {
         $this->http->mock
             ->when()
                 ->methodIs('POST')
             ->then()
-                ->callback(static function(Response $response) {$response->setContent('CALLBACK');})
+                ->callback(static function(Response $response): void {$response->setContent('CALLBACK');})
             ->end();
         $this->http->setUp();
-        $this->assertSame('CALLBACK', $this->http->client->post('/')->send()->getBody(true));
+        self::assertSame(
+            'CALLBACK',
+            (string) $this->http->client->sendRequest(
+                $this->getRequestFactory()->createRequest('POST', '/')
+            )->getBody()
+        );
     }
 
-    public function testComplexResponse()
+    public function testComplexResponse(): void
     {
         $this->http->mock
             ->when()
@@ -197,15 +211,20 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->body('BODY')
             ->end();
         $this->http->setUp();
-        $response = $this->http->client
-            ->post('/', ['x-client-header' => 'header-value'], ['post-key' => 'post-value'])->send();
-        $this->assertSame('BODY', $response->getBody(true));
-        $this->assertSame(201, $response->getStatusCode());
-        $this->assertSame('Bar', (string) $response->getHeader('X-Foo'));
-        $this->assertSame('post-value', $this->http->requests->latest()->getPostField('post-key'));
+        $response = $this->http->client->sendRequest(
+            $this->getRequestFactory()
+                ->createRequest('POST', '/')
+                ->withHeader('Content-Type', 'application/x-www-form-urlencoded')
+                ->withHeader('x-client-header', 'header-value')
+                ->withBody($this->getStreamFactory()->createStream(http_build_query(['post-key' => 'post-value'])))
+        );
+        self::assertSame('BODY', (string) $response->getBody());
+        self::assertSame(201, $response->getStatusCode());
+        self::assertSame('Bar', $response->getHeaderLine('X-Foo'));
+        self::assertSame('post-value', $this->http->requests->latest()->request->get('post-key'));
     }
 
-    public function testPutRequest()
+    public function testPutRequest(): void
     {
         $this->http->mock
             ->when()
@@ -216,15 +235,20 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->header('X-Foo', 'Bar')
             ->end();
         $this->http->setUp();
-        $response = $this->http->client
-            ->put('/', ['x-client-header' => 'header-value'], ['put-key' => 'put-value'])->send();
-        $this->assertSame('BODY', $response->getBody(true));
-        $this->assertSame(201, $response->getStatusCode());
-        $this->assertSame('Bar', (string) $response->getHeader('X-Foo'));
-        $this->assertSame('put-value', $this->http->requests->latest()->getPostField('put-key'));
+        $response = $this->http->client->sendRequest(
+            $this->getRequestFactory()
+                ->createRequest('PUT', '/')
+                ->withHeader('Content-Type', 'application/x-www-form-urlencoded')
+                ->withHeader('x-client-header', 'header-value')
+                ->withBody($this->getStreamFactory()->createStream(http_build_query(['put-key' => 'put-value'])))
+        );
+        self::assertSame('BODY', (string) $response->getBody());
+        self::assertSame(201, $response->getStatusCode());
+        self::assertSame('Bar', $response->getHeaderLine('X-Foo'));
+        self::assertSame('put-value', $this->http->requests->latest()->request->get('put-key'));
     }
 
-    public function testPostRequest()
+    public function testPostRequest(): void
     {
         $this->http->mock
             ->when()
@@ -235,15 +259,20 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
                 ->header('X-Foo', 'Bar')
             ->end();
         $this->http->setUp();
-        $response = $this->http->client
-            ->post('/', ['x-client-header' => 'header-value'], ['post-key' => 'post-value'])->send();
-        $this->assertSame('BODY', $response->getBody(true));
-        $this->assertSame(201, $response->getStatusCode());
-        $this->assertSame('Bar', (string) $response->getHeader('X-Foo'));
-        $this->assertSame('post-value', $this->http->requests->latest()->getPostField('post-key'));
+        $response = $this->http->client->sendRequest(
+            $this->getRequestFactory()
+                ->createRequest('POST', '/')
+                ->withHeader('Content-Type', 'application/x-www-form-urlencoded')
+                ->withHeader('x-client-header', 'header-value')
+                ->withBody($this->getStreamFactory()->createStream(http_build_query(['post-key' => 'post-value'])))
+        );
+        self::assertSame('BODY', (string) $response->getBody());
+        self::assertSame(201, $response->getStatusCode());
+        self::assertSame('Bar', $response->getHeaderLine('X-Foo'));
+        self::assertSame('post-value', $this->http->requests->latest()->request->get('post-key'));
     }
 
-    public function testCountRequests()
+    public function testCountRequests(): void
     {
         $this->http->mock
             ->when()
@@ -253,17 +282,22 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
             ->end();
         $this->http->setUp();
 
-        $this->assertCount(0, $this->http->requests);
-        $this->assertSame('resource body', (string) $this->http->client->get('/resource')->send()->getBody());
-        $this->assertCount(1, $this->http->requests);
+        self::assertCount(0, $this->http->requests);
+        self::assertSame(
+            'resource body',
+            (string) $this->http->client->sendRequest(
+                $this->getRequestFactory()->createRequest('GET', '/resource')
+            )->getBody()
+        );
+        self::assertCount(1, $this->http->requests);
     }
 
-    public function testMatchQueryString()
+    public function testMatchQueryString(): void
     {
         $this->http->mock
             ->when()
                 ->callback(
-                    function (Request $request) {
+                    static function (Request $request) {
                         return $request->query->has('key1');
                     }
                 )
@@ -273,13 +307,21 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
             ->end();
         $this->http->setUp();
 
-        $this->assertSame('query string', (string) $this->http->client->get('/?key1=')->send()->getBody());
-
-        $this->assertSame(Response::HTTP_NOT_FOUND, $this->http->client->get('/')->send()->getStatusCode());
-        $this->assertSame(Response::HTTP_NOT_FOUND, $this->http->client->post('/')->send()->getStatusCode());
+        self::assertSame(
+            'query string', (string) $this->http->client->sendRequest(
+            $this->getRequestFactory()->createRequest('GET', '/?key1='))->getBody()
+        );
+        self::assertSame(
+            Response::HTTP_NOT_FOUND,
+            $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/'))->getStatusCode()
+        );
+        self::assertSame(
+            Response::HTTP_NOT_FOUND,
+            $this->http->client->sendRequest($this->getRequestFactory()->createRequest('POST', '/'))->getStatusCode()
+        );
     }
 
-    public function testMatchRegex()
+    public function testMatchRegex(): void
     {
         $this->http->mock
             ->when()
@@ -289,11 +331,17 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
             ->end();
         $this->http->setUp();
 
-        $this->assertSame('response', (string) $this->http->client->get('/')->send()->getBody());
-        $this->assertSame('response', (string) $this->http->client->get('/')->send()->getBody());
+        self::assertSame(
+            'response',
+            (string) $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/'))->getBody()
+        );
+        self::assertSame(
+            'response',
+            (string) $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/'))->getBody()
+        );
     }
 
-    public function testMatchQueryParams()
+    public function testMatchQueryParams(): void
     {
         $this->http->mock
             ->when()
@@ -308,28 +356,31 @@ class HttpMockPHPUnitIntegrationTest extends AbstractTestCase
             ->end();
         $this->http->setUp();
 
-        $this->assertSame(
+        self::assertSame(
             'response',
-            (string) $this->http->client->get('/?p1=&p2=v2&p4=any&p5=v5&p6=v6')->send()->getBody()
-        );
-        $this->assertSame(
+            (string) $this->http->client->sendRequest(
+                $this->getRequestFactory()->createRequest('GET', '/?p1=&p2=v2&p4=any&p5=v5&p6=v6'))->getBody()
+            );
+        self::assertSame(
             Response::HTTP_NOT_FOUND,
-            $this->http->client->get('/?p1=&p2=v2&p3=foo')->send()->getStatusCode()
+            $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/?p1=&p2=v2&p3=foo'))
+                ->getStatusCode()
         );
-        $this->assertSame(
+        self::assertSame(
             Response::HTTP_NOT_FOUND,
-            $this->http->client->get('/?p1=')->send()->getStatusCode()
+            $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/?p1='))->getStatusCode()
         );
-        $this->assertSame(
+        self::assertSame(
             Response::HTTP_NOT_FOUND,
-            $this->http->client->get('/?p3=foo')->send()->getStatusCode()
+            $this->http->client->sendRequest($this->getRequestFactory()->createRequest('GET', '/?p3=foo'))
+                ->getStatusCode()
         );
     }
 
-    public function testFatalError()
+    public function testFatalError(): void
     {
-        if (version_compare(PHP_VERSION, '7.0', '<')) {
-            $this->markTestSkipped('Comment in to test if fatal errors are properly handled');
+        if (PHP_VERSION_ID < 70000) {
+            self::markTestSkipped('Comment in to test if fatal errors are properly handled');
         }
 
         $this->expectException('Error');
